Add gaming quotes
Unfortunately, all the trailing spaces need to be there, so that the
line breaks happen. You're weird, Markdown.

Even weirder: for lines ending in a dot dot dot double-quote (...")
character sequence, even two trailing spaces does not force a line
break. But if there are four dots, all is well. So that's why.